Lots of progress on debugging the Go-based compiler and JS runtime
There are several material changes in the internals here:
- argc is now the argument count, not including the selector
  (eg. `foo bar` is 0, `1 + 2` is 1,
  `send some: 7 values: 9 here: 'yo'` is 3).
- Primitives always return to the calling frame with some answer, they
  end the method in which they appear. There can be a preamble but no
  followup. If you need that, put the primitive in an internal method
  and call it from the larger method.
- Bugfixes and adjustments in the bytecodes as well.

Most importantly, we've moved from top-level >> (not a standard method
at all, just a human notation) to attach methods to a pair of actual
methods: One to create the `CompiledMethod` with two sends:
  CompiledMethod argc: c locals: n length: len.
  theClass method: (the above) selector: #whatever:selector:
These are sent to actual methods that are defined in the native kernel
to call primitive operations.

This more Smalltalk-y bootstrapping will simplify future in-universe
compiles and editing.